shallot: sweep extras
Rows consumed: animation (rename core.ts to playable.ts; owns and
getEasingIndex stay internal), cells, gltf, skin, text (merge core.ts
into index.ts), cells strip (grid.ts fillCellGrid with its memoized
pipeline, select.ts dispatchSelect), extras barrel (export * to explicit
names). Refused: orbit strip (examples ascii and render-to-a-terminal
enable OrbitTuning). Extras core subpaths point at index.ts until the
exports pass.
