+++ package.json
@@ -54,11 +54,11 @@
         "./physics/core": "./src/standard/physics/index.ts",
         "./character/core": "./src/standard/character/index.ts",
         "./bvh/core": "./src/standard/bvh/index.ts",
-        "./cells/core": "./src/extras/cells/core.ts",
-        "./gltf/core": "./src/extras/gltf/core.ts",
-        "./skin/core": "./src/extras/skin/core.ts",
-        "./text/core": "./src/extras/text/core.ts",
-        "./animation/core": "./src/extras/animation/core.ts",
+        "./cells/core": "./src/extras/cells/index.ts",
+        "./gltf/core": "./src/extras/gltf/index.ts",
+        "./skin/core": "./src/extras/skin/index.ts",
+        "./text/core": "./src/extras/text/index.ts",
+        "./animation/core": "./src/extras/animation/index.ts",
         "./runtime": "./src/engine/runtime/index.ts"
     },
     "bin": {
