+++ src/extras/index.ts
@@ -1,12 +1,97 @@
-// the extras barrel star-exports every module — each module's index.ts owns its clean public API.
-export * from "./animation";
-export * from "./cells";
-export * from "./gltf";
-export * from "./lines";
-export * from "./orbit";
-export * from "./outline";
-export * from "./profile";
-export * from "./skin";
-export * from "./sky";
-export * from "./sprite";
-export * from "./text";
+// the extras barrel lists each module's author names explicitly; extension names stay on the module.
+export {
+    AnimationPlugin,
+    AnimationState,
+    AnimationSystem,
+    Animator,
+    Composite,
+    type Easing,
+    Fill,
+    type Keyframe,
+    type KeyframeOptions,
+    keyframes,
+    mixer,
+    type Playable,
+    type PlayableEntry,
+    Playables,
+    Pose,
+    type Strip,
+    script,
+    type Track,
+} from "./animation";
+export {
+    CellsPlugin,
+    COLS,
+    cells,
+    cellsGridFor,
+    DEFAULT_FONT,
+    ROWS,
+} from "./cells";
+export {
+    type GltfHandle,
+    type GltfImport,
+    type GltfPlacement,
+    GltfPlugin,
+    loadGltf,
+    placeGltf,
+    placeScene,
+    Textured,
+} from "./gltf";
+export {
+    Arrow,
+    arrow,
+    box,
+    Line,
+    LinesPlugin,
+    segment,
+} from "./lines";
+export {
+    Orbit,
+    OrbitMode,
+    OrbitOverlayPlugin,
+    OrbitPick,
+    OrbitPlugin,
+    OrbitTuningPlugin,
+} from "./orbit";
+export {
+    Outline,
+    OutlinePlugin,
+} from "./outline";
+export {
+    type BenchmarkAPI,
+    type BenchmarkCompileStats,
+    type BenchmarkCpuStats,
+    type BenchmarkFrameStats,
+    type BenchmarkGpuStats,
+    type BenchmarkMeasurement,
+    type BenchmarkMemoryStats,
+    Profile,
+    ProfilePlugin,
+    showProfiler,
+} from "./profile";
+export {
+    LiveSkin,
+    LiveSkinSystem,
+    Skin,
+    SkinPlugin,
+    skinMatrix,
+    skinTraits,
+} from "./skin";
+export {
+    Sky,
+    SkyPlugin,
+} from "./sky";
+export {
+    image,
+    Sprite,
+    SpriteBillboard,
+    SpriteBlend,
+    SpriteFill,
+    SpritePlugin,
+} from "./sprite";
+export {
+    font,
+    Text,
+    TextPlugin,
+    text,
+} from "./text";
